+++ src/rules/function-comma-newline-after/__tests__/index.js
@@ -15,6 +15,7 @@ testRule("always", tr => {
   tr.ok("a { transform: translate(1 ,\n1); }")
   tr.ok("a { transform: translate(1,\r\n1); }", "CRLF")
   tr.ok("a { transform: color(rgb(0 ,\n\t0,\n\t0) lightness(50%)); }")
+  tr.ok("a { background: linear-gradient(45deg,\n rgba(0,\n 0,\n 0,\n 1),\n red); }")
 
   tr.notOk("a { transform: translate(1,1); }", {
     message: messages.expectedAfter(),
@@ -46,6 +47,11 @@ testRule("always", tr => {
     line: 2,
     column: 4,
   })
+  tr.notOk("a { background: linear-gradient(45deg,\n rgba(0,\n 0, 0,\n 1),\n red); }", {
+    message: messages.expectedAfter(),
+    line: 3,
+    column: 3,
+  })
 })
 
 testRule("always-multi-line", tr => {
@@ -70,6 +76,16 @@ testRule("always-multi-line", tr => {
         inset 0 -10px 12px 0 #f00;
     }
   `)
+  tr.ok(`
+    .foo {
+      background-image:
+        repeating-linear-gradient(
+          -45deg,
+          transparent,
+          rgba(0, 0, 0, 1) 5px
+        );
+      }
+  `)
 
   tr.notOk("a { transform: color(rgb(0 , 0 ,\n0) lightness(50%)); }", {
     message: messages.expectedAfterMultiLine(),
@@ -80,6 +96,16 @@ testRule("always-multi-line", tr => {
     message: messages.expectedAfterMultiLine(),
     line: 2,
     column: 4,
+  })
+  tr.notOk("a { background-image: repeating-linear-gradient(\n-45deg,\ntransparent, rgba(0, 0, 0, 1) 5px\n);}", {
+    message: messages.expectedAfterMultiLine(),
+    line: 3,
+    column: 12,
+  })
+  tr.notOk("a { background: linear-gradient(45deg,rgba(0,\n0 ,\n 0 ,\n 1)); }", {
+    message: messages.expectedAfterMultiLine(),
+    line: 1,
+    column: 38,
   })
 })
 
@@ -96,6 +122,7 @@ testRule("never-multi-line", tr => {
   tr.ok("a { transform: translate(1,  1); }")
   tr.ok("a { transform: translate(1, 1); }")
   tr.ok("a { transform: translate(1,\t1); }")
+  tr.ok("a { background: linear-gradient(45deg\n,rgba(0, 0, 0, 1),red); }")
 
   tr.notOk("a { transform: color(rgb(0 ,0 ,\n0) lightness(50%)); }", {
     message: messages.rejectedAfterMultiLine(),
@@ -116,5 +143,10 @@ testRule("never-multi-line", tr => {
     message: messages.rejectedAfterMultiLine(),
     line: 1,
     column: 43,
+  })
+  tr.notOk("a { background: linear-gradient(45deg\n,rgba(0,0 , 0, 1), red); }", {
+    message: messages.rejectedAfterMultiLine(),
+    line: 2,
+    column: 18,
   })
 })